Registro/app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0

google-cloud-firestore==2.19.0

argon2-cffi==23.1.0
PyJWT==2.10.1
pydantic==2.10.4
httpx==0.27.0
